Changes to matlab package and CMake code
-Started removing unneeded project() calls in CMake
-Started changing PROJECT_BINARY_DIR to CMAKE_CURRENT_BINARY_DIR
-Changed interface to crudeMCLib, and associated changes in executables /
matlab package / R package
-Added conditionalMC matlab interface
-Added dependencies for matlab shared libraries in CMake
-Added proper exception catching and memory cleanup in matlab

+++ splittingBasic/splittingBasic.cpp
@@ -132,7 +132,7 @@ namespace discreteGermGrain
 				float splittingFactor = splittingFactors[initialRadius-radius-1];
 				for(std::vector<::discreteGermGrain::subObs::basic>::iterator subObsIterator = observations.begin(); subObsIterator != observations.end(); subObsIterator++)
 				{
-					int integerSplittingFactor = splittingFactor;
+					int integerSplittingFactor = (int)splittingFactor;
 					if(bernoullis[initialRadius-radius-1](randomSource)) integerSplittingFactor++;
 					for(int k = 0; k < integerSplittingFactor; k++)
 					{
